Weak hash tables: Ensure ambiguous references to parts
This prevents a very subtle bug: The MPS AWL pool doesn't handle
ambiguous interior pointers.  The way DOHASH_WEAK is written, no base
pointer to the weak or strong parts of the weak hash table is on the
stack, so it's possible for the AWL segment not to update forwarding
references before allowing access to the segment.

* src/fns.c (hash_table_being_accessed): New variable.
(strengthen_hash_table): Ensure the strong and weak parts are pinned
by an ambiguous non-interior reference while we access the hash table.