Don't trace the pointer in Lisp_Misc_Ptr objects
We used to trace the pointer in Lisp_Misc_Ptr objects just in case it
points to an object on the GC heap without knowing much of the intended
use.  However, make_misc_ptr is primarily used by make_mint_ptr for
unaligned pointers.  MPS ignores unaligned references and, if enabled,
unaligned references cause an igc_assertion to fail.  So tracing the
pointer is either useless or counterproductive.

* src/igc.c (fix_misc_ptr): Ignore the pointer field.