vc-do-command: Error when BUFFER is nil, untabify
This function is shortly to gain support for separating standard
output and standard error from the child.  Commit untabifying it
first to make the diff adding the new features more readable.
This will include documented support for BUFFER nil, meaning to
discard output.

* lisp/vc/vc-dispatcher.el (vc-do-command): Untabify the source
in preparation for work on this function.  If BUFFER is nil,
signal an error, in keeping with the current docstring.
Use ensure-list.  Use equal not string= for string comparison.