Eglot: debounce consecutive semtok full/delta requests (bug#79374)
Many back-to-back calls for 'eglot--semtok-request' and small
regions occur even on trivial fast edits.  Even though it's
cheap and harmless to send many delta rquests, it's nicer to
just send just one.

The debouncing from jsonrpc-async-request isn't enough.  When a response
to a semtok request is received, it's important to remember which
regions to flush.  Font-lock logic will then do its magic and coalesce
those regions into one before invoking eglot--semtok-font-lock, and
eglot--semtok-font-lock-1 will have up-to-date data to work with.

* lisp/progmodes/eglot.el (eglot--semtok-inflight): New variable.
(eglot--semtok-request): Use it.