Eglot: debounce multiple useless full-buffer semtok delta rquests
Many back-to-back calls for 'eglot--semtok-request' and small
regions occur even on trivial fast edits.  Even though it's
cheap and harmless to send many delta rquests, it's nicer to
just send just one.  The debouncing from jsonrpc-async-request
almost suffices, but sometimes two requests creep in because
didChange is sent in between.  Use this simple flag to avoid.

* lisp/progmodes/eglot.el (eglot--semtok-inflight): New variable.
(eglot--semtok-request): Use it.